// src/store/inspire/mutation-types.ts
export enum MutationTypes {
  SetUserInvitationCodes = 'SET_USER_INVITATION_CODES',
  AppendUserInvitationCode = 'APPEND_USER_INVITATION_CODE',

  SetAppCommissionSetting = 'SET_APP_COMMISSION_SETTING',

  SetAppPurchaseAmountSettings = 'SET_APP_PURCHASE_AMOUNT_SETTINGS',
  AppendAppPurchaseAmountSetting = 'APPEND_APP_PURCHASE_AMOUNT_SETTING',

  SetCommissionCoins = 'SET_COMMISSION_COINS'
}